Move input handler functions out of main.go into inputhandlers.go
